[DB] frfrf

+++ Database/01_Implementazione/01_CREATE/12_JAVA_function.sql
@@ -5208,7 +5208,7 @@ BEGIN
         goal_conceded = new_goal_conceded::dm_usint,
         penalty_saved = new_penalty_saved::dm_usint
     WHERE
-        play_id = id_play;
+        play_id = id_play::integer;
 
     
     IF (0 = count_row) THEN
